+++ index.js
@@ -1,26 +1,20 @@
 'use strict'
 /* global fetch */
 
-const baseUrl = 'https://www.cryptocompare.com/api/data/'
+const baseUrl = 'https://min-api.cryptocompare.com/data/'
 
 function fetchJSON (url) {
   return fetch(url)
     .then(res => res.json())
     .then(body => {
-      // 'response' is a CryptoCompare field
-      if (body.Response !== 'Success') throw new Error(body.Message)
-      return body.Data
+      if (body.Response === 'Error') throw body.Message
+      return body
     })
 }
 
-function coinList () {
-  let url = `${baseUrl}coinlist/`
-  return fetchJSON(url)
-}
-
-function price (fsym, tsyms, useBtc) {
+function price (fsym, tsyms, tryConversion) {
   let url = `${baseUrl}price?fsym=${fsym}&tsyms=${tsyms}`
-  if (useBtc) url += 'usebtc=true'
+  if (tryConversion === false) url += '&tryConversion=false'
   return fetchJSON(url)
 }
 
@@ -28,11 +22,11 @@ function priceHistorical (fsym, tsyms, time) {
   if (!(time instanceof Date)) throw new Error('time parameter must be an instance of Date.')
   time = Math.floor(time.getTime() / 1000)
   let url = `${baseUrl}pricehistorical?fsym=${fsym}&tsyms=${tsyms}&ts=${time}`
-  return fetchJSON(url)
+  // The API returns json with an extra layer of nesting, so remove it
+  return fetchJSON(url).then(result => result[fsym])
 }
 
 module.exports = {
-  coinList,
   price,
   priceHistorical
 }